Seperate metadata seeding job from `ClassSeeder`
# Create `MetadataSeeder
# Apply to `DBInitializer
