+++ src/app/db/ClassSeeder.ts
@@ -1,5 +1,3 @@
-import { GlobalNameDoc } from 'Database/schemas/global-name'
-import metadataSeoulJSON from 'Resources/metadata/cau-seoul.json'
 import classesSeoulUnderJSON from 'Resources/classes/cau-seoul-학부.json'
 import classesSeoulGradJSON from 'Resources/classes/cau-seoul-대학원.json'
 import University from 'Database/models/university'
@@ -21,18 +19,6 @@ interface ClassesJSON {
   classes: ClassDoc[]
 }
 
-interface BldgMetaJSON {
-  number: string
-  name: string
-}
-
-interface UnivMetaJSON {
-  name: GlobalNameDoc
-  campus: GlobalNameDoc
-  vendor: string
-  buildings: BldgMetaJSON[]
-}
-
 // Set the current semester with this.
 const currentSemester = {
   year: '2019',
@@ -40,11 +26,6 @@ const currentSemester = {
 }
 
 export default class ClassSeeder {
-  /**
-   * Seed(resource) data for metadata
-   */
-  private metadataSeed: UnivMetaJSON[] = [metadataSeoulJSON as UnivMetaJSON]
-
   /**
    * Seed(resource) data for classes
    */
@@ -57,49 +38,13 @@ export default class ClassSeeder {
    * Seed the data associated with class.
    */
   public async run(): Promise<void> {
-    const univCount = await University.estimatedDocumentCount()
+    const classCount = await Class.estimatedDocumentCount()
 
-    // if vacant collection is empty, then seed data.
-    if (univCount === 0) {
-      await this.seedMetadata()
+    // if class collection is empty, then seed data.
+    if (classCount === 0) {
       await this.seedClasses()
       await this.linkClassesOfCurrentSemester()
     }
-  }
-
-  /**
-   * Seed the collection of university and building.
-   */
-  private async seedMetadata(): Promise<void> {
-    // asynchronously seed metadata
-    await Promise.all(
-      this.metadataSeed.map(async (seed: UnivMetaJSON) => {
-        // create university
-        const university = await University.create({
-          name: seed.name,
-          campus: seed.campus,
-          vendor: seed.vendor
-        })
-
-        // create buildings
-        const buildings = await Building.insertMany(seed.buildings)
-        // get building id list
-        const buildingsIds = buildings.map(bldg => {
-          return bldg._id
-        })
-
-        // link university to buildings
-        await Building.updateMany(
-          { _id: { $in: buildingsIds } },
-          { university: university._id }
-        )
-
-        // link buildings to university
-        await University.findByIdAndUpdate(university._id, {
-          buildings: buildings
-        })
-      })
-    )
   }
 
   /**
@@ -234,27 +179,5 @@ export default class ClassSeeder {
           '` are unregistered.'
       )
     }
-  }
-
-  /**
-   * Apply refreshed seed data.
-   *
-   * @deprecated
-   */
-  public async refreshMetadataAndClasses(): Promise<void> {
-    // delete all documents about metadata and classes
-    await Promise.all([
-      Class.deleteMany({}),
-      University.deleteMany({}),
-      Building.deleteMany({}),
-      Floor.deleteMany({}),
-      Classroom.deleteMany({}),
-      Lecture.deleteMany({})
-    ])
-
-    // re-seeding
-    await this.seedMetadata()
-    await this.seedClasses()
-    await this.linkClassesOfCurrentSemester()
   }
 }
