+++ src/app/db/DBInitializer.ts
@@ -1,13 +1,17 @@
 import ClassSeeder from 'DB/ClassSeeder'
-import EmptyCount from 'Database/models/empty-count'
 import MealSeeder from 'DB/MealSeeder'
+import MetadataSeeder from 'DB/MetadataSeeder'
+import EmptyCount from 'Database/models/empty-count'
 
 export default class DBInitializer {
+  private metadataSeeder: MetadataSeeder
+
   private classSeeder: ClassSeeder
 
   private mealSeeder: MealSeeder
 
   public constructor() {
+    this.metadataSeeder = new MetadataSeeder()
     this.classSeeder = new ClassSeeder()
     this.mealSeeder = new MealSeeder()
   }
@@ -16,19 +20,21 @@ export default class DBInitializer {
    * Initialize database.
    */
   public async initialize(): Promise<void> {
+    await this.metadataSeeder.run()
+
     const promises = []
     promises.push(this.classSeeder.run())
     promises.push(this.mealSeeder.run())
-
     await Promise.all(promises) // asynchronously seeding
-    await this.calcEmptyCounts()
+
+    await this.calculateEmptyCounts()
   }
 
   /**
    * Delete all previous empty classroom counts.
    * Calculate and save empty classroom counts in advance.
    */
-  private async calcEmptyCounts(): Promise<void> {
+  private async calculateEmptyCounts(): Promise<void> {
     await EmptyCount.deletePrevCounts()
     await EmptyCount.saveCurrentCount()
     await EmptyCount.saveNextCount()
